move GetBlockInfo to common
